Split into two services, one a fake for generating HL7 messages, and one for converting them into interchange messages. Messages are not genuine HL7 and are not yet sent.
