+++ waveform-reader/src/main/java/uk/ac/ucl/rits/inform/datasources/waveform/Hl7ListenerConfig.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 import org.springframework.integration.annotation.ServiceActivator;
 import org.springframework.integration.ip.tcp.TcpReceivingChannelAdapter;
 import org.springframework.integration.ip.tcp.connection.DefaultTcpNetConnectionSupport;
@@ -21,8 +22,15 @@ import java.util.List;
  * Listen on a TCP port for incoming HL7 messages.
  */
 @Configuration
+@Profile("hl7reader")
 public class Hl7ListenerConfig {
     private final Logger logger = LoggerFactory.getLogger(Hl7ListenerConfig.class);
+
+    private final Hl7ParseAndSend hl7ParseAndSend;
+
+    public Hl7ListenerConfig(Hl7ParseAndSend hl7ParseAndSend) {
+        this.hl7ParseAndSend = hl7ParseAndSend;
+    }
 
     /**
      * Specify the server config.
@@ -73,11 +81,14 @@ public class Hl7ListenerConfig {
     /**
      * Message handler. Source IP check has passed if we get here. No reply is expected.
      * @param msg the incoming message
+     * @throws InterruptedException .
      */
     @ServiceActivator(inputChannel = "hl7Stream")
-    public void handler(Message<?> msg) {
+    public void handler(Message<?> msg) throws InterruptedException {
         byte[] asBytes = (byte[]) msg.getPayload();
         String asStr = new String(asBytes, StandardCharsets.UTF_8);
+        // XXX: parse message from HL7 to interchange message, send to publisher
+        hl7ParseAndSend.parseAndSend(asStr);
         logger.info("MESSAGE {}", asStr);
     }
 }
